+++ cloudsmith_cli/cli/commands/__init__.py
@@ -6,6 +6,7 @@ from . import check  # noqa
 from . import delete  # noqa
 from . import docs  # noqa
 from . import help_  # noqa
+from . import list_  # noqa
 from . import push  # noqa
 from . import status  # noqa
 from . import token  # noqa